users_backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
mysqlclient==2.2.1
djangorestframework_simplejwt==5.3.1
requests==2.31.0
cryptography==41.0.7